requirements.txt
easydict==1.9
numpy==1.19.1
PyYAML~=5.3.1
scipy==1.5.4
torch==1.9.1+cu102
torchvision==0.10.1+cu102
tqdm==4.56.0

Pillow~=8.3.2
opencv-python~=4.5.4.58
loguru~=0.6.0
tabulate~=0.8.9
matplotlib~=3.4.3
pycocotools~=2.0.3